+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_data.py
@@ -2,6 +2,7 @@ import math
 from dataclasses import dataclass
 from datetime import date
 from typing import Any
+
 import numpy as np
 
 from projects._03_independent_cqc._02_clean.fargate.utils.ascwds_filled_posts_calculator.difference_within_range import (
@@ -17,6 +18,7 @@ from utils.column_values.categorical_column_values import (
     Dormancy,
     MainJobRoleLabels,
     PrimaryServiceType,
+    Region,
     Sector,
 )
 
@@ -472,6 +474,40 @@ class ValidateCleanIndCQCData:
         (date(2024, 1, 1), "1-002", Sector.independent, "N", None),
         (date(2024, 2, 1), "1-001", Sector.independent, "Y", 10),
         (date(2024, 2, 1), "1-002", Sector.independent, "N", None),
+    ] # fmt: skip
+
+
+@dataclass
+class ValidateImputedIndCqcAscwdsAndPir:
+    cleaned_ind_cqc_rows = [
+        ("1-000000001", date(2024, 1, 1),),
+        ("1-000000002", date(2024, 1, 1),),
+        ("1-000000001", date(2024, 2, 1),),
+        ("1-000000002", date(2024, 2, 1),),
+    ] # fmt: skip
+
+    imputed_ind_cqc_ascwds_and_pir_rows = [
+        ("1-000000001", date(2024, 1, 1), date(2024, 1, 1), date(2024, 1, 1), "Y", "prov_1", Sector.independent, date(2024, 1, 1), "Y", 5, PrimaryServiceType.care_home_only, date(2024, 1, 1), "cssr", "region", date(2024, 1, 1), "cssr", "region", "RUI", "lsoa", "msoa", 5, 5, "ascwds_filtering_rule", "source", 5.0, 5, 123456789, 5, 5.0),
+        ("1-000000002", date(2024, 1, 1), date(2024, 1, 1), date(2024, 1, 1), "Y", "prov_1", Sector.independent, date(2024, 1, 1), "Y", 5, PrimaryServiceType.care_home_only, date(2024, 1, 1), "cssr", "region", date(2024, 1, 1), "cssr", "region", "RUI", "lsoa", "msoa", 5, 5, "ascwds_filtering_rule", "source", 5.0, 5, 123456789, 5, 5.0),
+        ("1-000000001", date(2024, 1, 9), date(2024, 1, 1), date(2024, 1, 1), "Y", "prov_1", Sector.independent, date(2024, 1, 1), "Y", 5, PrimaryServiceType.care_home_only, date(2024, 1, 1), "cssr", "region", date(2024, 1, 1), "cssr", "region", "RUI", "lsoa", "msoa", 5, 5, "ascwds_filtering_rule", "source", 5.0, 5, 123456789, 5, 5.0),
+        ("1-000000002", date(2024, 1, 9), date(2024, 1, 1), date(2024, 1, 1), "Y", "prov_1", Sector.independent, date(2024, 1, 1), "Y", 5, PrimaryServiceType.care_home_only, date(2024, 1, 1), "cssr", "region", date(2024, 1, 1), "cssr", "region", "RUI", "lsoa", "msoa", 5, 5, "ascwds_filtering_rule", "source", 5.0, 5, 123456789, 5, 5.0),
+    ] # fmt: skip
+
+
+@dataclass
+class ValidateEstimatedIndCQCFilledPostsData:
+    imputed_ind_cqc_rows = [
+        ("1-000000001", date(2024, 1, 1)),
+        ("1-000000002", date(2024, 1, 1)),
+        ("1-000000001", date(2024, 2, 1)),
+        ("1-000000002", date(2024, 2, 1)),
+    ]
+
+    estimated_ind_cqc_filled_posts_rows = [
+        ("1-000000001", date(2024, 1, 1), date(2024, 1, 1), "Y", Sector.independent, 5, PrimaryServiceType.care_home_only, PrimaryServiceType.care_home_only, date(2024, 1, 1), "cssr", "region", 5, 5, 5, "source", 5.0, 5.0, 5, 5.0, 123456789, 5.0, "source", 5.0, 5.0, 5.0, 5.0, 5.0, 5.0),
+        ("1-000000002", date(2024, 1, 1), date(2024, 1, 1), "Y", Sector.independent, 5, PrimaryServiceType.care_home_only, PrimaryServiceType.care_home_only, date(2024, 1, 1), "cssr", "region", 5, 5, 5, "source", 5.0, 5.0, 5, 5.0, 123456789, 5.0, "source", 5.0, 5.0, 5.0, 5.0, 5.0, 5.0),
+        ("1-000000001", date(2024, 1, 9), date(2024, 1, 1), "Y", Sector.independent, 5, PrimaryServiceType.care_home_only, PrimaryServiceType.care_home_only, date(2024, 1, 1), "cssr", "region", 5, 5, 5, "source", 5.0, 5.0, 5, 5.0, 123456789, 5.0, "source", 5.0, 5.0, 5.0, 5.0, 5.0, 5.0),
+        ("1-000000002", date(2024, 1, 9), date(2024, 1, 1), "Y", Sector.independent, 5, PrimaryServiceType.care_home_only, PrimaryServiceType.care_home_only, date(2024, 1, 1), "cssr", "region", 5, 5, 5, "source", 5.0, 5.0, 5, 5.0, 123456789, 5.0, "source", 5.0, 5.0, 5.0, 5.0, 5.0, 5.0),
     ] # fmt: skip
 
 
@@ -1716,4 +1752,57 @@ class OutlierCleaningData:
         ("1-004", 40),
         ("1-004", 45),
         ("1-004", 50),
+    ]
+
+
+@dataclass
+class EstimateFilledPostsModelsUtils:
+    enrich_model_ind_cqc_rows = [
+        ("1-001", date(2025, 1, 1), CareHome.not_care_home, None),
+        ("1-002", date(2025, 1, 1), CareHome.not_care_home, None),
+        ("1-003", date(2025, 1, 1), CareHome.care_home, 2),
+        ("1-004", date(2025, 1, 1), CareHome.care_home, 2),
+    ]
+
+    enrich_model_predictions_care_home_rows = [
+        ("1-003", date(2025, 1, 1), 2, -0.5, "v1_r1"),
+        ("1-004", date(2025, 1, 1), 2, 2.5, "v1_r1"),
+    ]
+
+    expected_enrich_model_ind_cqc_care_home_rows = [
+        ("1-001", date(2025, 1, 1), CareHome.not_care_home, None, None, None),  # no prediction expected
+        ("1-002", date(2025, 1, 1), CareHome.not_care_home, None, None, None),  # no prediction expected
+        ("1-003", date(2025, 1, 1), CareHome.care_home, 2, -1.0, "v1_r1"),  # prediction (converted to posts) joined in (maintains negative)
+        ("1-004", date(2025, 1, 1), CareHome.care_home, 2, 5.0, "v1_r1"),  # prediction (converted to posts) joined in
+    ] # fmt: skip
+
+    enrich_model_predictions_non_res_rows = [
+        ("1-001", date(2025, 1, 1), 2, -5.0, "v1_r1"),
+        ("1-002", date(2025, 1, 1), 2, 2.5, "v1_r1"),
+    ]
+
+    expected_enrich_model_ind_cqc_non_res_rows = [
+        ("1-001", date(2025, 1, 1), CareHome.not_care_home, None, -5.0, "v1_r1"),  # prediction joined in (maintains negative)
+        ("1-002", date(2025, 1, 1), CareHome.not_care_home, None, 2.5, "v1_r1"),  # prediction joined in
+        ("1-003", date(2025, 1, 1), CareHome.care_home, 2, None, None),  # no prediction expected
+        ("1-004", date(2025, 1, 1), CareHome.care_home, 2, None, None),  # no prediction expected
+    ] # fmt: skip
+    join_ind_cqc_rows = [
+        ("1-001", Region.london, 67, date(2022, 2, 20)),
+        ("1-001", Region.london, 67, date(2022, 3, 29)),
+        ("1-002", Region.north_east, 12, date(2022, 3, 29)),
+    ]
+
+    join_prediction_rows = [
+        ("1-001", 67, date(2022, 3, 29), 10.0, "v1.0.0_r2"),
+    ]
+    expected_join_without_run_id_rows = [
+        ("1-001", Region.london, 67, date(2022, 2, 20), None),
+        ("1-001", Region.london, 67, date(2022, 3, 29), 10.0),
+        ("1-002", Region.north_east, 12, date(2022, 3, 29), None),
+    ]
+    expected_join_with_run_id_rows = [
+        ("1-001", Region.london, 67, date(2022, 2, 20), None, None),
+        ("1-001", Region.london, 67, date(2022, 3, 29), 10.0, "v1.0.0_r2"),
+        ("1-002", Region.north_east, 12, date(2022, 3, 29), None, None),
     ]